pyproject.toml
[project]
name = "recipe-telegram-bot"
version = "0.1.0"
description = "A telegram bot that communicated with smart recipe generator"
authors = [
    {name = "Shahzaib Ali Khan",email = "shahzaib.ali.k39@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot (>=22.5,<23.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
]

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "isort (>=6.1.0,<7.0.0)",
    "black (>=25.9.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
